Delete person commands
