+++ src/main/java/seedu/address/logic/commands/ClearTransactionsCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.transaction;
+package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
