+++ src/main/java/seedu/address/logic/commands/GetTotalExpenditureCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.transaction;
+package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_CATEGORY;
@@ -8,8 +8,6 @@ import java.time.Month;
 import seedu.address.commons.enums.TransactionType;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.UniCashMessages;
-import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.transaction.Category;
