--- src/main/java/seedu/address/logic/commands/transaction/DeleteTransactionCommand.java
@@ -1,72 +1,0 @@
-package seedu.address.logic.commands.transaction;
-
-import static java.util.Objects.requireNonNull;
-
-import java.util.List;
-
-import seedu.address.commons.core.index.Index;
-import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.logic.UniCashMessages;
-import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.CommandResult;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
-import seedu.address.model.transaction.Transaction;
-
-/**
- * This class handles the deletion of a Transaction from the UniCa$h database.
- */
-public class DeleteTransactionCommand extends Command {
-
-    public static final String COMMAND_WORD = "delete_transaction";
-
-    public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Deletes the transaction identified by the index number used in the displayed transaction list.\n"
-            + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
-
-    public static final String MESSAGE_DELETE_TRANSACTION_SUCCESS = "Deleted Transaction: %1$s";
-
-    private final Index targetIndex;
-
-    public DeleteTransactionCommand(Index targetIndex) {
-        this.targetIndex = targetIndex;
-    }
-
-    @Override
-    public CommandResult execute(Model model) throws CommandException {
-        requireNonNull(model);
-        List<Transaction> lastShownList = model.getFilteredTransactionList();
-
-        if (targetIndex.getZeroBased() >= lastShownList.size()) {
-            throw new CommandException(UniCashMessages.MESSAGE_INVALID_TRANSACTION_DISPLAYED_INDEX);
-        }
-
-        Transaction transactionToDelete = lastShownList.get(targetIndex.getZeroBased());
-        model.deleteTransaction(transactionToDelete);
-        return new CommandResult(String.format(MESSAGE_DELETE_TRANSACTION_SUCCESS,
-                UniCashMessages.formatTransaction(transactionToDelete)));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof DeleteTransactionCommand)) {
-            return false;
-        }
-
-        DeleteTransactionCommand otherDeleteCommand = (DeleteTransactionCommand) other;
-        return targetIndex.equals(otherDeleteCommand.targetIndex);
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .add("targetIndex", targetIndex)
-                .toString();
-    }
-}
