+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -8,20 +8,16 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.transaction.AddTransactionCommand;
-import seedu.address.logic.commands.transaction.ClearTransactionsCommand;
-import seedu.address.logic.commands.transaction.DeleteTransactionCommand;
-import seedu.address.logic.commands.transaction.EditTransactionCommand;
-import seedu.address.logic.commands.transaction.GetTotalExpenditureCommand;
+import seedu.address.logic.commands.AddTransactionCommand;
+import seedu.address.logic.commands.ClearTransactionsCommand;
+import seedu.address.logic.commands.DeleteTransactionCommand;
+import seedu.address.logic.commands.EditTransactionCommand;
+import seedu.address.logic.commands.GetTotalExpenditureCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.logic.parser.transaction.AddTransactionCommandParser;
 import seedu.address.logic.parser.transaction.DeleteTransactionCommandParser;
