+++ src/main/java/seedu/address/logic/parser/transaction/DeleteTransactionCommandParser.java
@@ -2,7 +2,7 @@ package seedu.address.logic.parser.transaction;
 import static seedu.address.logic.UniCashMessages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.transaction.DeleteTransactionCommand;
+import seedu.address.logic.commands.DeleteTransactionCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
