+++ src/test/java/seedu/address/logic/commands/transaction/AddTransactionCommandTest.java
@@ -18,6 +18,7 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.UniCashMessages;
+import seedu.address.logic.commands.AddTransactionCommand;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
