+++ src/test/java/seedu/address/logic/commands/transaction/ClearTransactionsCommandTest.java
@@ -5,6 +5,7 @@ import static seedu.address.testutil.TypicalTransactions.getTypicalUniCash;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.ClearTransactionsCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
