+++ src/test/java/seedu/address/logic/commands/transaction/GetTotalExpenditureCommandTest.java
@@ -8,7 +8,8 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.enums.TransactionType;
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.ClearTransactionsCommand;
+import seedu.address.logic.commands.GetTotalExpenditureCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
