+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -14,20 +14,15 @@ import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
-import seedu.address.logic.commands.transaction.AddTransactionCommand;
-import seedu.address.logic.commands.transaction.ClearTransactionsCommand;
-import seedu.address.logic.commands.transaction.DeleteTransactionCommand;
-import seedu.address.logic.commands.transaction.EditTransactionCommand;
-import seedu.address.logic.commands.transaction.GetTotalExpenditureCommand;
+import seedu.address.logic.commands.AddTransactionCommand;
+import seedu.address.logic.commands.ClearTransactionsCommand;
+import seedu.address.logic.commands.DeleteTransactionCommand;
+import seedu.address.logic.commands.EditTransactionCommand;
+import seedu.address.logic.commands.GetTotalExpenditureCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Person;
 import seedu.address.model.transaction.Transaction;
