+++ src/test/java/seedu/address/logic/parser/transaction/DeleteTransactionCommandParserTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.transaction.DeleteTransactionCommand;
+import seedu.address.logic.commands.DeleteTransactionCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 public class DeleteTransactionCommandParserTest {
